Speech-App/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
azure-cognitiveservices-speech==1.41.1
